requirements.txt

openpyxl
pillow
sqlalchemy>=2
# -- DB Drivers --
asyncpg
python-dotenv
# -- Web Server --
uvicorn
hypercorn
numpy==1.26.4
scipy==1.10.1
pandas==2.1.4
scikit-learn==1.3.2
matplotlib==3.8.2
geopandas==0.14.2
shapely==2.0.2
quart==0.19.4
imageio
imageio-ffmpeg
werkzeug

